+++ CHANGELOG.md
@@ -7,10 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## Changed
+### Added
+- Config option to add additional containers to be clicked to.
+
+### Changed
+- Empty Fast Item Frames can now be clicked through.
+
+## [3.4.1] - 2025-04-16
+
+### Fixed
+- Crash on Forge.
+
+## [3.4] -2025-04-11
+
+### Changed
 - Ported buildscript to Stonecutter. Mod functionality should now be more consistent across versions.
 
-## Fixed
+### Fixed
 - Configs are now saved when the game is closed, fixing an issue where the config wouldn't be saved when changed via the keybind.
 
 ## [3.3] - 2025-03-09